pyproject.toml
[project]
name = "smart-home-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomqtt>=2.4.0",
    "asgi-correlation-id>=4.3.4",
    "fastapi>=0.116.1",
    "granian>=2.5.1",
    "homeassistant-api>=4.2.2.post1",
    "langchain-core>=0.3.74",
    "langsmith[openai-agents]>=0.4.16",
    "line-bot-sdk>=3.18.1",
    "openai-agents>=0.2.9",
    "pydantic-settings>=2.10.1",
    "pydantic-settings-yaml>=0.2.0",
    "structlog>=25.4.0",
    "temporalio>=1.16.0",
]
